detaching "Find" functionnality
