--- src/com/pad/noob/INoobPad.java
@@ -1,7 +1,0 @@
-package com.pad.noob;
-
-public interface INoobPad {
-	void newFile();
-	void findWindow();
-
-}
